search by category done

+++ client/src/components/ArticleDetail.vue
@@ -8,15 +8,18 @@
     <div class="row">
       <div class="col-md-12">
         <p class="content" v-for="content in article.articleContent.split('\n')">{{ content }}</p>
+        <p>by {{ article.author.username }}</p>
       </div>
     </div>
     <div class="row">
       <div class="col-10">
       </div>
-      <div class="col-2">
-        <i class="fa fa-trash buttonme" aria-hidden="true" @click="doRemove"></i>
+      <div class="col-2" v-if="user._id === article.author">
+        <i class="fa fa-trash buttonme" aria-hidden="true" @click="doRemove" style="color: red"></i>
         <i class="fa fa-pencil-square-o buttonme" aria-hidden="true" @click="editArticle"></i>
+
       </div>
+    </div>
     </div>
   </div>
 </template>
@@ -39,15 +42,25 @@ export default {
     },
     doRemove () {
       var self = this
-      let id = this.$route.query.id
-      this.axios.delete(`http://localhost:3000/api/articles/${id}`, {
-        token: localStorage.getItem('token')
-      })
-      .then(response => {
-        console.log(response)
-        self.$store.dispatch('getArticles')
-        self.$router.push('/')
-      })
+      let id = self.$route.query.id
+      if (window.confirm('Are you sure?')) {
+        this.axios.delete(`http://localhost:3000/api/articles/${id}`, {
+          token: localStorage.getItem('token')
+        })
+        .then(response => {
+          console.log(response)
+          self.$store.dispatch('getArticles')
+          self.$router.push('/')
+        })
+      }
+    },
+    editArticle () {
+
+    }
+  },
+  computed: {
+    user () {
+      return this.$store.getters.user
     }
   },
   created () {
